requirements.txt
google-generativeai>=0.3.0
pandas>=2.0.0
openpyxl>=3.1.0